Make volsync safer in cases where the nas is not working

+++ k8s-volsync/volsync.sh
@@ -7,12 +7,23 @@ cd ${VOLROOT} || exit
 PERIOD="2h"
 if [ -n "$1" ]; then PERIOD="$1";fi
 
+# Function to check if a volume is completely empty (ignoring lost+found)
+is_empty() {
+  [ -z "$(ls -A "$1" | grep -v 'lost+found')" ]
+}
+
 function handle_sigterm() {
   echo SIGTERM received - shutting down
   for VOL in */*; do
-    echo "Syncing $VOL to tank"
-    echo -n "tank" >"${VOLROOT}/${VOL}/.nodeName"
-    rsync -avx -e "ssh ${SSH_OPTS}" --delete-during --exclude "lost+found" "${VOLROOT}/${VOL}/" "root@nas.default:/tank/Volumes/${VOL}/"
+    if [ -d "$VOL" ]; then
+      if is_empty "$VOL"; then
+        echo "WARNING: ${VOL} is completely empty! Skipping sync to prevent backup erasure."
+        continue
+      fi
+      echo "Syncing $VOL to tank"
+      echo -n "tank" >"${VOLROOT}/${VOL}/.nodeName"
+      rsync -avx -e "ssh ${SSH_OPTS}" --delete-during --exclude "lost+found" "${VOLROOT}/${VOL}/" "root@nas.default:/tank/Volumes/${VOL}/"
+    fi
   done
   echo Shutdown Complete
   exit
@@ -28,9 +39,15 @@ while true; do
   sleep "$PERIOD" &
   wait $!
   for VOL in */*; do
-    echo "Syncing $VOL to tank"
-    echo -n "tank" >"${VOLROOT}/${VOL}/.nodeName"
-    ionice -c 3 rsync -avx -e "ssh ${SSH_OPTS}" --delete-during --exclude "lost+found" "${VOLROOT}/${VOL}/" "root@nas.default:/tank/Volumes/${VOL}/"
+    if [ -d "$VOL" ]; then
+      if is_empty "$VOL"; then
+        echo "WARNING: ${VOL} is completely empty! Skipping sync to prevent backup erasure."
+        continue
+      fi
+      echo "Syncing $VOL to tank"
+      echo -n "tank" >"${VOLROOT}/${VOL}/.nodeName"
+      ionice -c 3 rsync -avx -e "ssh ${SSH_OPTS}" --delete-during --exclude "lost+found" "${VOLROOT}/${VOL}/" "root@nas.default:/tank/Volumes/${VOL}/"
+    fi
   done
   echo Sync Complete
 done
